token generator script(v7)/requirements.txt
absl-py==0.14.1
aiohttp==3.5.4
aiosonic==0.13.0
alabaster==0.7.12
aniso8601==9.0.1
apprise==0.9.2
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bidict==0.21.4
bleach==4.1.0
blis==0.7.4
bs4==0.0.1
bump2version==1.0.1
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.6
cheroot==8.5.2
chinese-converter==1.0.2
clang==5.0
click==8.0.3
colorama==0.4.4
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
crypto==1.4.1
cryptography==35.0.0
cycler==0.10.0
cymem==2.0.5
dateparser==1.1.0
decorator==5.1.0
discord==1.7.3
discord.py==1.0.1
distlib==0.3.3
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
eventlet==0.30.2
ffmpy==0.3.0
filelock==3.3.0
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.0.0
goto-statement==1.2
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hanziconv==0.3.2
hex==2.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.3
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
jaraco.functools==3.3.0
Jinja2==3.0.2
JsonForm==0.0.2
jsonschema==4.2.1
JsonSir==0.0.2
keras==2.6.0
keras-nightly==2.8.0.dev2021101007
Keras-Preprocessing==1.1.2
keyring==23.2.1
kiwisolver==1.3.2
lameenc==1.3.1
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
mido==1.2.10
more-itertools==8.10.0
multidict==4.7.6
murmurhash==1.0.5
Naked==0.1.31
numpy==1.19.5
oauthlib==3.1.1
onecache==0.1.0
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pathlib==1.0.1
pathy==0.6.0
pbkdf2==1.3
pexpect==4.8.0
Pillow==8.3.2
pipreqs==0.4.11
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
preshed==3.0.5
pretty-midi==0.2.9
protobuf==3.18.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.11.0
pycryptodomex==3.11.0
pydantic==1.8.2
pyfiglet==0.8.post1
pyflakes==2.2.0
Pygments==2.10.0
pyparsing==2.4.7
pypianoroll==1.0.4
pyrsistent==0.18.0
pytest==6.2.2
pytest-runner==5.2
python-binance==1.0.12
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
python-engineio==4.3.0
python-resources==0.3
python-socketio==5.2.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==30.0
regex==2021.10.23
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
Resource==0.2.1
rfc3986==1.5.0
rsa==4.7.2
savify==2.3.4
schedule==1.1.0
scipy==1.7.1
sha1==1.2
shellescape==3.8.1
six==1.15.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.3.1
spacy==3.1.3
spacy-legacy==3.0.8
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spotipy==2.19.0
SQLAlchemy==1.4.15
sqlitedict==1.7.0
srsly==2.4.1
Strip==0.0.1.dev14
tasksio==0.0.0
tb-nightly==2.7.0a20211010
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021101008
tf-nightly==2.8.0.dev20211010
thinc==8.0.10
tldextract==3.1.2
toml==0.10.2
tox==3.21.3
tqdm==4.62.3
translation==1.0.5
twine==3.3.0
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
unicorn-binance-websocket-api==1.30.0
unicorn-fy==0.11.0
uritemplate==4.1.1
urllib3==1.26.7
ushlex==0.99.1
utils==1.0.1
validators==0.18.2
virtualenv==20.8.1
wasabi==0.8.2
watchdog==1.0.2
webencodings==0.5.1
websocket-client==1.2.1
websockets==6.0
Werkzeug==2.0.2
wrapt==1.12.1
yarg==0.1.9
yarl==1.7.2
youtube-dl==2021.6.6
zipp==3.6.0
